main edits

+++ src/client/js/app.js
@@ -11,9 +11,8 @@ document.getElementById("add-trip-button").addEventListener("click", () => {
       return response.json();
     })
     .then((data) => {
-      const id = null;
-      // These variables are specific to each trip
-      const tripDateInput = document.getElementById("trip-date").value;
+      const tripDateStart = document.getElementById("trip-date").value;
+      const tripDateEnd = document.getElementById("end-date").value;
       const countryName = data?.geonames[0]?.countryName;
       const location_latitude = data?.geonames[0]?.lat;
       const location_longitude = data?.geonames[0]?.lng;
@@ -22,12 +21,13 @@ document.getElementById("add-trip-button").addEventListener("click", () => {
         country: countryName,
         latitude: location_latitude,
         longitude: location_longitude,
-        date: tripDateInput,
+        date: tripDateStart,
+        endDate: tripDateEnd,
       };
       const saved = false;
-      updateUI(trip, saved);
+      const id = trip.id;
+      updateUI(trip, saved, id);
     })
-
     .catch((error) => {
       console.error(
         "There has been a problem with your fetch operation:",
@@ -47,16 +47,18 @@ document.addEventListener("DOMContentLoaded", () => {
     .then((data) => {
       data.forEach((trip) => {
         const saved = true;
-        updateUI(trip, saved);
+        const id = trip.id;
+        updateUI(trip, saved, id);
       });
     })
     .catch((error) => {
       console.error("Error fetching trips:", error);
     });
 });
+
 async function fetchImageURL(city, country) {
   const pixabayKey = process.env.pixabayKey;
-  const imageFetchURL = `https://pixabay.com/api/?key=${pixabayKey}&q=tourist+placis+${city}+in+${country}&image_type=photo`;
+  const imageFetchURL = `https://pixabay.com/api/?key=${pixabayKey}&q=Tourist+places+in+${city}+${country}&image_type=photo`;
   try {
     const response = await fetch(imageFetchURL);
     if (!response.ok) {
@@ -91,19 +93,21 @@ async function fetchWeatherStatus(long, lat) {
     return null;
   }
 }
-async function updateUI(trip, saved) {
+
+async function updateUI(trip, saved, id) {
   console.log(trip);
   const tripCard = document.createElement("div");
   tripCard.className = "trip-card";
   const tripDate = new Date(trip.date);
+  const endDate = new Date(trip.endDate);
   const currentDate = new Date();
-  const timeDiff = tripDate - currentDate;
-  const daysAway = Math.ceil(timeDiff / (1000 * 60 * 60 * 24));
+  const daysAway = Client.calculateDaysDiff(currentDate, trip.date);
+
+  const triplength = Client.calculateDaysDiff(trip.date, trip.endDate);
 
   // Fetch image URL
   const imageurl = await fetchImageURL(trip.city, trip.country);
   const weather = await fetchWeatherStatus(trip.longitude, trip.latitude);
-
   tripCard.innerHTML = `
     <img
       src="${imageurl}"
@@ -112,44 +116,62 @@ async function updateUI(trip, saved) {
     />
     <div class="trip-details">
       <p><strong>My trip to:</strong> ${trip.city}, ${trip.country}</p>
-      <p><strong>Departing:</strong> ${trip.date}</p>
+      <p><strong>Departing:</strong> ${trip.date} <strong>to:</strong> ${
+    trip.endDate
+  }</p>
+      <p><strong>Trip length:</strong> ${triplength} days</p>
       <div class="trip-buttons">
-      ${
-        saved
-          ? `<button class="remove-trip">Remove Trip</button>`
-          : `<button class="save-trip">Save Trip</button>`
-      }
+        ${
+          saved
+            ? `<button class="remove-trip">Remove Trip</button>`
+            : `<button class="save-trip">Save Trip</button>`
+        }
       </div>
       <p class="trip-info">${trip.city}, ${
     trip.country
   } is ${daysAway} days away</p>
-    ${
-      daysAway < 7 && daysAway > 0
-        ? `<p class="trip-weather">
+   ${
+     daysAway < 7 && daysAway > 0
+       ? `<p class="trip-weather">
         Typical weather in ${weather?.data[daysAway]?.datetime} is:<br />
         max tempreture - ${weather?.data[daysAway]?.app_max_temp} min temp - ${weather?.data[daysAway]?.app_min_temp}<br />
         ${weather?.data[daysAway]?.weather?.description}
       </p>`
-        : `<p>whether appear for 7 days or less`
-    }
-    
+       : `<p>whether appear for 7 days or less`
+   }
+      <div class="list">
+        <div class="list-item" id="add-item">
+          <p>Add to-do item<span>+</span></p>
+        </div>
+        ${
+          trip.list
+            ? trip.list
+                .map((item) => `<div class="list-item"><p>${item}</p></div>`)
+                .join("")
+            : ""
+        }
+       
+      </div>
     </div>
   `;
 
-  // Append the trip card to the container
   document.getElementById("travel-cont").appendChild(tripCard);
 
-  // Add event listener for saving the trip
-  tripCard.querySelector(".save-trip")?.addEventListener("click", (event) => {
+  tripCard.querySelector(".save-trip")?.addEventListener("click", () => {
+    const listItems = Array.from(tripCard.querySelectorAll(".list-item p"))
+      .filter((item, index) => index > 0) // Skip the first item
+      .map((item) => item.textContent);
     const tripData = {
       city: trip.city,
       country: trip.country,
       latitude: trip.latitude,
       longitude: trip.longitude,
       date: trip.date,
+      endDate: trip.endDate,
+      list: listItems,
     };
+    console.log(tripData);
 
-    // Send trip data to the server
     fetch("http://localhost:3000/add", {
       method: "POST",
       headers: {
@@ -160,6 +182,7 @@ async function updateUI(trip, saved) {
       .then((response) => response.json())
       .then((data) => {
         console.log("Trip saved:", data);
+        id = data.id;
         tripCard.querySelector(".trip-buttons").innerHTML = `
           <button class="remove-trip">Remove Trip</button>
         `;
@@ -170,15 +193,34 @@ async function updateUI(trip, saved) {
       });
   });
 
-  // Add event listener for removing the trip
+  tripCard.querySelector(".list-item").addEventListener("click", function () {
+    let newItemText = prompt("Enter a new to-do item:");
+    if (newItemText) {
+      let newListItem = document.createElement("div");
+      newListItem.className = "list-item";
+      let newP = document.createElement("p");
+      newP.textContent = newItemText;
+      newListItem.appendChild(newP);
+      tripCard.querySelector(".list").appendChild(newListItem);
+      if (id != null) {
+        tripCard.querySelector(".trip-buttons").innerHTML = `
+        <button class="update-list">update Trip</button>
+      `;
+        updateTrip(tripCard, trip, id);
+      }
+    }
+  });
+
   if (saved) {
     addRemoveTripListener(tripCard, trip.id);
   }
 }
 
-// Function to add remove trip event listener
 function addRemoveTripListener(tripCard, tripId) {
-  tripCard.querySelector(".remove-trip").addEventListener("click", (event) => {
+  if (!tripId) {
+    throw new Error("tripId must be found");
+  }
+  tripCard.querySelector(".remove-trip").addEventListener("click", () => {
     fetch(`http://localhost:3000/delete/${tripId}`, {
       method: "DELETE",
     })
@@ -192,5 +234,36 @@ function addRemoveTripListener(tripCard, tripId) {
       .catch((error) => {
         console.error("Error removing trip:", error);
       });
+  });
+}
+function updateTrip(tripCard, trip, id) {
+  tripCard.querySelector(".update-list").addEventListener("click", async () => {
+    const listItems = Array.from(tripCard.querySelectorAll(".list-item p"))
+      .filter((item, index) => index > 0) // Skip the first item
+      .map((item) => item.textContent);
+    const updatedTripData = { ...trip, list: listItems };
+
+    try {
+      const response = await fetch(`http://localhost:3000/update/${id}`, {
+        method: "PUT",
+        headers: {
+          "Content-Type": "application/json",
+        },
+        body: JSON.stringify(updatedTripData),
+      });
+
+      if (!response.ok) {
+        throw new Error("Network response was not ok " + response.statusText);
+      }
+
+      const data = await response.json();
+      tripCard.querySelector(".trip-buttons").innerHTML = `
+      <button class="remove-trip">Remove Trip</button>
+    `;
+      addRemoveTripListener(tripCard, data.id);
+      console.log("Trip updated:", data);
+    } catch (error) {
+      console.error("Error updating trip:", error);
+    }
   });
 }
